Refactor the user dropdown menu into its own component
