+++ scss/appNav.module.scss
@@ -6,9 +6,11 @@
     height: 65px;
   }
 }
+
 .navbar {
   font-family: PT Mono;
 }
+
 .navbar-brand {
   font-size: 40px;
 }
@@ -34,23 +36,6 @@
   }
 }
 
-.nav-user-toggle {
-  height: 38px;
-  width: 38px;
-  display: block;
-  text-decoration: none;
-  color: #000;
-  &::after {
-    display: none;
-  }
-}
-
-.down-icon {
-  position: relative;
-  bottom: 30px;
-  left: 42px;
-}
-
 .nav-buttons {
   flex-shrink: 0;
   & > * {
@@ -71,27 +56,9 @@
   color: #5440d8;
 }
 
-.dropdown-item {
-  color: #565656;
-  text-decoration: none;
-  font-style: normal;
-  font-weight: 700;
-  text-transform: uppercase;
-  font-size: 14px;
-  &:hover {
-    background-color: color.change(variables.$primary, $lightness: 90%);
-    color: #4c4c4c;
-    text-decoration: none;
-  }
-}
-
 .active {
   background: #5440d8 !important;
   color: #ffffff !important;
-}
-
-.dropdown-menu {
-  box-shadow: 0px 4px 25px 0px hsla(0, 0%, 0%, 0.07);
 }
 
 .light-button > * {
