+++ scss/dropDown.module.scss
@@ -1,21 +1,45 @@
+@use 'sass:color';
+@use '_variables.module.scss';
 @import 'colors.module.scss';
 
-.item {
-  color: $gray;
+.nav-user-toggle {
+  display: flex;
+  text-decoration: none;
+  color: #343a40;
+  align-items: center;
+  grid-column-gap: 8px;
+  cursor: pointer;
+  text-transform: capitalize;
+  font-family: Inter;
+  &::after {
+    display: none;
+  }
+}
+
+.dropdown-item {
+  color: #565656;
+  text-decoration: none;
+  font-style: normal;
+  font-weight: 700;
+  text-transform: uppercase;
+  font-size: 14px;
   &:hover {
-    color: $darkGray;
+    background-color: color.change(variables.$primary, $lightness: 90%);
+    color: #4c4c4c;
     text-decoration: none;
   }
 }
 
-.title {
-  color: $gray;
-  background: transparent;
-  border: none;
-  padding: 0.5em;
-  margin: 0 2em;
-  &:hover {
-    color: $darkGray;
-    text-decoration: none;
-  }
+.active {
+  background: #5440d8 !important;
+  color: #ffffff !important;
+}
+
+.user-icon {
+  height: 20px;
+  width: 20px;
+}
+
+.dropdown-menu {
+  box-shadow: 0px 4px 25px 0px hsla(0, 0%, 0%, 0.07);
 }
